@@ -18,44 +18,41 @@ use Jobcloud\Kafka\SchemaRegistryClient\Exception\SubjectNotFoundException;
 use Jobcloud\Kafka\SchemaRegistryClient\Exception\UnauthorizedException;
 use Jobcloud\Kafka\SchemaRegistryClient\Exception\UnprocessableEntityException;
 use Jobcloud\Kafka\SchemaRegistryClient\Exception\VersionNotFoundException;
+use PHPUnit\Framework\Attributes\CoversClass;
+use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 use Psr\Http\Message\RequestInterface;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\StreamInterface;
 
-/**
- * @covers \Jobcloud\Kafka\SchemaRegistryClient\ErrorHandler
- */
+#[CoversClass(ErrorHandler::class)]
 class ErrorHandlerTest extends TestCase
 {
-    private const TEST_MESSAGE = 'Test Message';
+    private const string TEST_MESSAGE = 'Test Message';
 
-    private function makeResponseInterfaceMock(?string $code = null, ?string $message = null): MockObject
+    private function makeResponseInterfaceMock(?int $code = null, ?string $message = null): MockObject
     {
-        $streamMock = $this
-            ->getMockBuilder(StreamInterface::class)
-            ->onlyMethods(['__toString'])
-            ->getMockForAbstractClass();
+        $streamMock = $this->createMock(StreamInterface::class);
 
         $streamMock
+            ->expects(self::once())
             ->method('__toString')
             ->willReturn(json_encode([
                 'error_code' => $code,
                 'message' => $message,
             ]));
 
-        $responseMock = $this
-            ->getMockBuilder(ResponseInterface::class)
-            ->onlyMethods(['getBody'])
-            ->getMockForAbstractClass();
+        $responseMock = $this->createMock(ResponseInterface::class);
 
-        $responseMock->method('getBody')->willReturn($streamMock);
+        $responseMock->expects(self::once())
+            ->method('getBody')
+            ->willReturn($streamMock);
 
         return $responseMock;
     }
 
-    public function exceptionTestDataProvider(): array
+    public static function exceptionTestDataProvider(): array
     {
         return [
             [50001, BackendDatastoreException::class],
@@ -82,8 +79,6 @@ class ErrorHandlerTest extends TestCase
     }
 
     /**
-     * @dataProvider exceptionTestDataProvider
-     *
      * @throws BackendDatastoreException
      * @throws ClientException
      * @throws CompatibilityException
@@ -100,6 +95,7 @@ class ErrorHandlerTest extends TestCase
      * @throws UnprocessableEntityException
      * @throws VersionNotFoundException
      */
+    #[DataProvider('exceptionTestDataProvider')]
     public function testExceptionThrow(?int $code, string $expectedException): void
     {
         /** @var ResponseInterface|MockObject $responseMock */
@@ -130,10 +126,17 @@ class ErrorHandlerTest extends TestCase
     {
         /** @var ResponseInterface|MockObject $responseMock */
         $responseMock = $this->makeResponseInterfaceMock(50001, self::TEST_MESSAGE);
-        $requestMock = $this->getMockForAbstractClass(RequestInterface::class);
-        $streamMock = $this->getMockForAbstractClass(StreamInterface::class);
-        $streamMock->expects(self::once())->method('getContents')->willReturn('test body');
-        $requestMock->expects(self::once())->method('getBody')->willReturn($streamMock);
+
+        $requestMock = $this->createMock(RequestInterface::class);
+        $streamMock = $this->createMock(StreamInterface::class);
+
+        $requestMock->expects(self::once())
+            ->method('getBody')
+            ->willReturn($streamMock);
+
+        $streamMock->expects(self::once())
+            ->method('getContents')
+            ->willReturn('test body');
 
         $errorHandler = new ErrorHandler();
 
