@@ -2,21 +2,22 @@
 
 namespace Jobcloud\Kafka\SchemaRegistryClient\Tests;
 
+use Buzz\Client\Curl;
 use Jobcloud\Kafka\SchemaRegistryClient\ErrorHandlerInterface;
 use Jobcloud\Kafka\SchemaRegistryClient\HttpClientInterface;
 use Jobcloud\Kafka\SchemaRegistryClient\KafkaSchemaRegistryApiClientInterface;
 use Jobcloud\Kafka\SchemaRegistryClient\ServiceProvider\KafkaSchemaRegistryApiClientProvider;
 use LogicException;
+use Nyholm\Psr7\Factory\Psr17Factory;
 use phpmock\phpunit\MockObjectProxy;
 use phpmock\phpunit\PHPMock;
+use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 use Pimple\Container;
 use Psr\Http\Message\RequestFactoryInterface;
 
-/**
- * @covers \Jobcloud\Kafka\SchemaRegistryClient\ServiceProvider\KafkaSchemaRegistryApiClientProvider
- */
+#[CoversClass(KafkaSchemaRegistryApiClientProvider::class)]
 class KafkaSchemaRegistryApiClientProviderTest extends TestCase
 {
     use PHPMock;
@@ -39,10 +40,15 @@ class KafkaSchemaRegistryApiClientProviderTest extends TestCase
 
     public function testDefaultContainersAndServicesSetWithMinimalConfig(): void
     {
-        $this->classExistsMock->expects(self::exactly(2))->withConsecutive(
-            ['Nyholm\Psr7\Factory\Psr17Factory'],
-            ['Buzz\Client\Curl']
-        )->willReturn(true);
+        $expectedCalls = [Psr17Factory::class, Curl::class];
+        $callIndex = 0;
+
+        $this->classExistsMock
+            ->expects(self::exactly(2))
+            ->willReturnCallback(function ($className) use ($expectedCalls, &$callIndex) {
+                self::assertSame($expectedCalls[$callIndex++], $className);
+                return true;
+            });
 
         $container = new Container();
 
@@ -86,13 +92,15 @@ class KafkaSchemaRegistryApiClientProviderTest extends TestCase
 
     public function testSuccessWithMissingAuth(): void
     {
+        $expectedCalls = [Psr17Factory::class, Curl::class];
+        $callIndex = 0;
+
         $this->classExistsMock
             ->expects(self::exactly(2))
-            ->withConsecutive(
-                ['Nyholm\Psr7\Factory\Psr17Factory'],
-                ['Buzz\Client\Curl']
-            )
-            ->willReturn(true);
+            ->willReturnCallback(function ($className) use ($expectedCalls, &$callIndex) {
+                self::assertSame($expectedCalls[$callIndex++], $className);
+                return true;
+            });
 
         $container = new Container();
 
@@ -128,10 +136,15 @@ class KafkaSchemaRegistryApiClientProviderTest extends TestCase
 
     public function testUserNameAndPasswordFromSettingsArePassedToHttpClient(): void
     {
-        $this->classExistsMock->expects(self::exactly(2))->withConsecutive(
-            ['Nyholm\Psr7\Factory\Psr17Factory'],
-            ['Buzz\Client\Curl']
-        )->willReturn(true);
+        $expectedCalls = [Psr17Factory::class, Curl::class];
+        $callIndex = 0;
+
+        $this->classExistsMock
+            ->expects(self::exactly(2))
+            ->willReturnCallback(function ($className) use ($expectedCalls, &$callIndex) {
+                self::assertSame($expectedCalls[$callIndex++], $className);
+                return true;
+            });
 
         $container = new Container();
 
