@@ -6,21 +6,20 @@ use Jobcloud\Kafka\SchemaRegistryClient\Exception\ImportException;
 use Jobcloud\Kafka\SchemaRegistryClient\Exception\SchemaNotFoundException;
 use Jobcloud\Kafka\SchemaRegistryClient\Exception\SubjectNotFoundException;
 use Jobcloud\Kafka\SchemaRegistryClient\Exception\VersionNotFoundException;
-use Jobcloud\Kafka\SchemaRegistryClient\HttpClient;
 use Jobcloud\Kafka\SchemaRegistryClient\HttpClientInterface;
 use Jobcloud\Kafka\SchemaRegistryClient\KafkaSchemaRegistryApiClient;
 use Jobcloud\Kafka\SchemaRegistryClient\KafkaSchemaRegistryApiClientInterface;
+use PHPUnit\Framework\Attributes\CoversClass;
+use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 
-/**
- * @covers \Jobcloud\Kafka\SchemaRegistryClient\KafkaSchemaRegistryApiClient
- */
+#[CoversClass(KafkaSchemaRegistryApiClient::class)]
 class KafkaSchemaRegistryApiClientTest extends TestCase
 {
-    private const TEST_SUBJECT_NAME = 'some-subject';
-    private const TEST_SCHEMA = '{}';
-    private const TEST_VERSION = 3;
+    private const string TEST_SUBJECT_NAME = 'some-subject';
+    private const string TEST_SCHEMA = '{}';
+    private const int TEST_VERSION = 3;
 
     public function testGetSubjects(): void
     {
@@ -125,9 +124,7 @@ class KafkaSchemaRegistryApiClientTest extends TestCase
         $api->getSchemaById(1);
     }
 
-    /**
-     * @dataProvider schemaDataProvider
-     **/
+    #[DataProvider('schemaDataProvider')]
     public function testRegisterNewSchemaVersion(string $testSchema, string $expectedSchema): void
     {
         $httpClientMock = $this->getHttpClientMock();
@@ -166,9 +163,7 @@ class KafkaSchemaRegistryApiClientTest extends TestCase
         self::assertFalse($result);
     }
 
-    /**
-     * @dataProvider schemaDataProvider
-     **/
+    #[DataProvider('schemaDataProvider')]
     public function testCheckSchemaCompatibilityForVersionTrue(string $testSchema, string $expectedSchema): void
     {
         $httpClientMock = $this->getHttpClientMock();
@@ -256,7 +251,7 @@ class KafkaSchemaRegistryApiClientTest extends TestCase
             ->willThrowException(new VersionNotFoundException());
 
         $api = new KafkaSchemaRegistryApiClient($httpClientMock);
-        $result = $api->checkSchemaCompatibilityForVersion(
+        $api->checkSchemaCompatibilityForVersion(
             self::TEST_SUBJECT_NAME,
             self::TEST_SCHEMA,
             self::TEST_VERSION
@@ -285,7 +280,6 @@ class KafkaSchemaRegistryApiClientTest extends TestCase
         $result = $api->checkSchemaCompatibilityForVersion(
             self::TEST_SUBJECT_NAME,
             self::TEST_SCHEMA,
-            KafkaSchemaRegistryApiClient::VERSION_LATEST
         );
         self::assertTrue($result);
     }
@@ -309,23 +303,40 @@ class KafkaSchemaRegistryApiClientTest extends TestCase
         self::assertSame(KafkaSchemaRegistryApiClientInterface::LEVEL_FULL, $result);
     }
 
-    public function testGetDefaultCompatibiltyLeveWhenGetSubjectCompatibilityLevelThrowsException(): void
+    public function testGetDefaultCompatibilityLevelWhenGetSubjectCompatibilityLevelThrowsException(): void
     {
         $httpClientMock = $this->getHttpClientMock();
 
-        $httpClientMock
+        $expectedCalls = [
+                [
+                    'method' => 'GET',
+                    'uri' => sprintf('config/%s', self::TEST_SUBJECT_NAME),
+                    'throw' => true,
+                ],
+                [
+                    'method' => 'GET',
+                    'uri' => 'config',
+                    'return' => [
+                        'compatibilityLevel' => KafkaSchemaRegistryApiClientInterface::LEVEL_FULL,
+                    ],
+                ],
+            ];
+            $callIndex = 0;
+
+            $httpClientMock
             ->expects(self::exactly(2))
             ->method('call')
-            ->withConsecutive(
-                ['GET', sprintf('config/%s', self::TEST_SUBJECT_NAME)],
-                []
-            )
-            ->will(
-                $this->onConsecutiveCalls(
-                    $this->throwException(new SubjectNotFoundException()),
-                    ['compatibilityLevel' => KafkaSchemaRegistryApiClientInterface::LEVEL_FULL]
-                )
-            );
+            ->willReturnCallback(function (string $method, string $uri) use ($expectedCalls, &$callIndex) {
+                $expected = $expectedCalls[$callIndex++];
+                self::assertSame($expected['method'], $method);
+                self::assertSame($expected['uri'], $uri);
+
+                if (isset($expected['throw'])) {
+                    throw new SubjectNotFoundException();
+                }
+
+                return $expected['return'];
+            });
 
         $api = new KafkaSchemaRegistryApiClient($httpClientMock);
         $result = $api->getSubjectCompatibilityLevel(self::TEST_SUBJECT_NAME);
@@ -348,8 +359,7 @@ class KafkaSchemaRegistryApiClientTest extends TestCase
 
         $api = new KafkaSchemaRegistryApiClient($httpClientMock);
         $result = $api->setSubjectCompatibilityLevel(
-            self::TEST_SUBJECT_NAME,
-            KafkaSchemaRegistryApiClientInterface::LEVEL_FULL
+            self::TEST_SUBJECT_NAME
         );
 
         self::assertTrue($result);
@@ -362,7 +372,7 @@ class KafkaSchemaRegistryApiClientTest extends TestCase
         $httpClientMock
             ->expects(self::once())
             ->method('call')
-            ->with('GET', sprintf('config'))
+            ->with('GET', 'config')
             ->willReturn(['compatibilityLevel' => KafkaSchemaRegistryApiClientInterface::LEVEL_FULL]);
 
         $api = new KafkaSchemaRegistryApiClient($httpClientMock);
@@ -386,9 +396,7 @@ class KafkaSchemaRegistryApiClientTest extends TestCase
         self::assertTrue($result);
     }
 
-    /**
-     * @dataProvider schemaDataProvider
-     **/
+    #[DataProvider('schemaDataProvider')]
     public function testGetVersionForSchema(string $testSchema, string $expectedSchema): void
     {
         $httpClientMock = $this->getHttpClientMock();
@@ -565,14 +573,10 @@ class KafkaSchemaRegistryApiClientTest extends TestCase
      */
     private function getHttpClientMock(): MockObject
     {
-        return $this
-            ->getMockBuilder(HttpClient::class)
-            ->disableOriginalConstructor()
-            ->onlyMethods(['call'])
-            ->getMock();
+        return $this->createMock(HttpClientInterface::class);
     }
 
-    public function schemaDataProvider(): array
+    public static function schemaDataProvider(): array
     {
         return [
             'empty schema' => ['{}', '[]'],
